add exceptionHandler

+++ application/src/main/java/com/senla/library/controller/AuthorController.java
@@ -1,5 +1,6 @@
 package com.senla.library.controller;
 import com.senla.library.dto.AuthorDTO;
+import com.senla.library.dto.UserDTO;
 import com.senla.library.exception.ResourceNotFoundException;
 import com.senla.library.service.AuthorService;
 import io.swagger.annotations.Api;
@@ -7,6 +8,7 @@ import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -35,22 +37,26 @@ public class AuthorController {
     @GetMapping("/author/{id}")
     public AuthorDTO findById(@PathVariable("id") Long id) {
         Optional<AuthorDTO> dtoOptional = authorService.findById(id);
-        return dtoOptional.orElseThrow(() -> new ResourceNotFoundException("User not found with id :" + id));
+        return dtoOptional.orElse(null);
     }
     
     @ApiOperation("Find by Id")
     @DeleteMapping("/author/delete/{id}")
-    public void delete(@PathVariable("id") Long id) {
-        authorService.deleteById(id);
-    
-    }
+    public ResponseEntity<AuthorDTO> deleteAuthor(@PathVariable ("id") Long id) {
+        Optional<AuthorDTO> authorDTO = authorService.findById(id);
+        if( authorDTO!= null){
+            authorService.deleteById(id);
+        }else throw new ResourceNotFoundException("User not found with id :" + id);
+            return ResponseEntity.ok().build();
+        }
+      
+        
     
     @ApiOperation("Find all data")
     @GetMapping("/list")
     public List<AuthorDTO> list() {
         return authorService.findAll();
     }
-    
     
     @ApiOperation("Pagination request")
     @GetMapping("/page-query")
